% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x)
}
\arguments{
\item{x}{A vector of data on which you wish to calculate the skewness}
}
\value{
A single skewness value
}
\description{
Skewness
}
\examples{
x <- df$values
skewness(x)

}
